backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0

pymongo==4.17.0
python-dotenv==1.2.2

numpy==1.26.4

opencv-python-headless==4.10.0.84

face-recognition==1.3.0
face-recognition-models==0.3.0
dlib==20.0.1

passlib==1.7.4
bcrypt==3.2.2

python-jose==3.5.0

cffi==1.17.1

starlette==0.46.0

click==8.3.3
idna==3.13
h11==0.16.0
anyio==4.13.0
typing_extensions==4.15.0
annotated-types==0.7.0
pydantic==2.13.3
pydantic_core==2.46.3